Group common styles into a single CSS selector
Styles that are common between icons (i.e. everything except the `content` property) are currently duplicated for each icon once the generated `.scss` file is converted to CSS.

This PR extracts these styles from the mixin and add them back via an `@extend` directive, which creates a combined selector for all of the icons. The `%icon` placeholder itself is never rendered in the final CSS.